requirements.txt

pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2


streamlit==1.38.0

matplotlib==3.9.2
seaborn==0.13.2

google-generativeai==0.1.3
python-dotenv==1.0.1